requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==0.62.1
django-debug-toolbar==4.3.0
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==9.5.0
typing_extensions==4.11.0
urllib3==2.2.1
